Fixing Review comments and indentations

+++ src/bp_sim.cpp
@@ -4533,7 +4533,7 @@ int CFlowGenList::load_astf(){
 
 int CFlowGenList::load_from_yaml(std::string file_name,
                                  uint32_t num_threads){
-    uint16_t idx;
+    uint8_t idx;
     m_yaml_info.load_from_yaml_file(file_name);
     if (m_yaml_info.verify_correctness(num_threads) ==false){
         exit(0);
@@ -5392,7 +5392,7 @@ rte_mbuf_t * CPluginCallbackSimple::sip_voice_plugin(uint8_t plugin_id,CGenNode 
 
                     // Mark as IPv6 and set the upper 96-bits
                     via_replace_cmd.m_flags |= CMiniVMCmdBase::MIN_VM_V6;
-                    for (uint16_t idx=0; idx<6; idx++){
+                    for (uint8_t idx=0; idx<6; idx++){
                         via_replace_cmd.m_ip.v6[idx] = CGlobalInfo::m_options.m_src_ipv6[idx];
                         via_replace_cmd.m_ip_via.v6[idx] = CGlobalInfo::m_options.m_src_ipv6[idx];
                     }
